Move benchmark parameter enums to main module
Summary:
I’d like to write a data analysis script to post-process the output of
the JMH benchmark. That script should live in the `main` module, but it
will also be convenient to reference the enum types used by the JMH
benchmark as its parameters. To avoid introducing a cyclic module
dependency, I’m moving these types to a container class in the main
module. This seems like a reasonable change in and of itself, too: it
significantly shrinks the contents and conceptual surface area of the
benchmark file, it removes the need to duplicate the resources across
modules, and it also opens the possibility of using the same
configuration types in the standalone benchmark.