Check in results
Summary:
The JMH benchmark takes a long time to run, so it only seems fair to
check these in for others’ convenience.

Each of these was computed after a fresh boot of my laptop from tty1,
with minimal other load. My laptop is a Lenovo T440s with an i5-4300U
(4-core) and 12 GB of RAM.